+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Carla.Build.cs
@@ -4,25 +4,17 @@ using System;
 using System.IO;
 using UnrealBuildTool;
 
-public class Carla : ModuleRules
+
+public class Carla : CarlaRules
 {
   bool UsingCarSim = false;
   bool UsingChrono = false;
   bool UsingPytorch = false;
-  bool UsingRos2 = false;
-  private bool IsWindows(ReadOnlyTargetRules Target)
-  {
-    return (Target.Platform == UnrealTargetPlatform.Win64) || (Target.Platform == UnrealTargetPlatform.Win32);
-  }
-
-  public Carla(ReadOnlyTargetRules Target) : base(Target)
+  public Carla(ReadOnlyTargetRules Target) : base(Target, "../../CarlaDependencies")
   {
     PrivatePCHHeaderFile = "Carla.h";
-
-    if (IsWindows(Target))
-    {
-      bEnableExceptions = true;
-    }
+    bEnableExceptions = true;
+    CppStandard = CppStandardVersion.Cpp20;
 
     // Read config about carsim
     string CarlaPluginPath = Path.GetFullPath( ModuleDirectory );
@@ -57,8 +49,12 @@ public class Carla : ModuleRules
       {
         Console.WriteLine("Enabling ros2");
         UsingRos2 = true;
-        PublicDefinitions.Add("WITH_ROS2");
-        PrivateDefinitions.Add("WITH_ROS2");
+      }
+
+      if (line.Contains("Debug ON"))
+      {
+        Console.WriteLine("Enabling debug build");
+        Debug = true;
       }
     }
 
@@ -91,10 +87,10 @@ public class Carla : ModuleRules
       PublicDependencyModuleNames.AddRange(new string[] { "CarSim" });
     }
 
-	 if (Target.Type == TargetType.Editor)
-	 {
-		PublicDependencyModuleNames.AddRange(new string[] { "UnrealEd" });
-	 }
+	  if (Target.Type == TargetType.Editor)
+	  {
+      PublicDependencyModuleNames.AddRange(new string[] { "UnrealEd" });
+	  }
 
     PrivateDependencyModuleNames.AddRange(
       new string[]
@@ -118,229 +114,143 @@ public class Carla : ModuleRules
         "PhysicsCore"
         // ... add private dependencies that you statically link with here ...
       }
-      );
+    );
+
     if (UsingCarSim)
     {
       PrivateDependencyModuleNames.AddRange(new string[] { "CarSim" });
       PrivateIncludePathModuleNames.AddRange(new string[] { "CarSim" });
     }
 
-
     DynamicallyLoadedModuleNames.AddRange(
       new string[]
       {
         // ... add any modules that your module loads dynamically here ...
       }
-      );
+    );
 
-    AddCarlaServerDependency(Target);
+    AddCarlaServerDependency();
   }
 
-  private bool UseDebugLibs(ReadOnlyTargetRules Target)
+  private void AddCarlaServerDependency()
   {
-    if (IsWindows(Target))
+
+    // Link dependencies.
+    AddStaticLibrary( "rpc");
+    if (UseDebugLibs())
     {
-      // In Windows, Unreal uses the Release C++ Runtime (CRT) even in debug
-      // mode, so unless we recompile the engine we cannot link the debug
-      // libraries.
-      return false;
+      AddStaticLibrary("carla_server_debug");
     }
     else
     {
-      return false;
+      AddStaticLibrary("carla_server");
     }
-  }
-
-  private void AddDynamicLibrary(string library)
-  {
-    PublicAdditionalLibraries.Add(library);
-    RuntimeDependencies.Add(library);
-    PublicDelayLoadDLLs.Add(library);
-  }
-  private void AddDllDependency(string PathToFolder, string DllName)
-  {
-    string Source = Path.Combine(PathToFolder, DllName);
-    string Destination = Path.Combine("$(BinaryOutputDir)", DllName);
-    RuntimeDependencies.Add(Destination, Source);
-  }
-
-  delegate string ADelegate(string s);
-
-  private void AddBoostLibs(string LibPath)
-  {
-    string [] files = Directory.GetFiles(LibPath, "*boost*.lib");
-    foreach (string file in files) PublicAdditionalLibraries.Add(file);
-  }
-
-  private void AddCarlaServerDependency(ReadOnlyTargetRules Target)
-  {
-    string LibCarlaInstallPath = Path.GetFullPath(Path.Combine(ModuleDirectory, "../../CarlaDependencies"));
-
-    ADelegate GetLibName = (string BaseName) => {
-      if (IsWindows(Target))
-      {
-        return BaseName + ".lib";
-      }
-      else
-      {
-        return "lib" + BaseName + ".a";
-      }
-    };
-
-    // Link dependencies.
-    if (IsWindows(Target))
+    if (IsWindows())
     {
-      AddBoostLibs(Path.Combine(LibCarlaInstallPath, "lib"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("rpc")));
+       PublicAdditionalLibraries.Add("shlwapi.lib");
 
-      if (UseDebugLibs(Target))
-      {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_server_debug")));
-      }
-      else
-      {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_server")));
-      }
+      AddBoostLibs();
       if (UsingChrono)
       {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("ChronoEngine")));
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("ChronoEngine_vehicle")));
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("ChronoModels_vehicle")));
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("ChronoModels_robot")));
+        AddStaticLibrary("ChronoEngine");
+        AddStaticLibrary("ChronoEngine_vehicle");
+        AddStaticLibrary("ChronoModels_vehicle");
+        AddStaticLibrary("ChronoModels_robot");
         AddDllDependency(Path.Combine(LibCarlaInstallPath, "dll"), "ChronoEngine.dll");
         AddDllDependency(Path.Combine(LibCarlaInstallPath, "dll"), "ChronoEngine_vehicle.dll");
         AddDllDependency(Path.Combine(LibCarlaInstallPath, "dll"), "ChronoModels_vehicle.dll");
         AddDllDependency(Path.Combine(LibCarlaInstallPath, "dll"), "ChronoModels_robot.dll");
         bUseRTTI = true;
       }
-
-      //OsmToODR
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "sqlite3.lib"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "xerces-c_3.lib"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "proj.lib"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "osm2odr.lib"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "zlibstatic.lib"));
     }
     else
     {
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("rpc")));
-      if (UseDebugLibs(Target))
-      {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_server_debug")));
-      }
-      else
-      {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_server")));
-      }
       if (UsingChrono)
       {
-        AddDynamicLibrary(Path.Combine(LibCarlaInstallPath, "lib", "libChronoEngine.so"));
-        AddDynamicLibrary(Path.Combine(LibCarlaInstallPath, "lib", "libChronoEngine_vehicle.so"));
-        AddDynamicLibrary(Path.Combine(LibCarlaInstallPath, "lib", "libChronoModels_vehicle.so"));
-        AddDynamicLibrary(Path.Combine(LibCarlaInstallPath, "lib", "libChronoModels_robot.so"));
+        AddStaticLibrary( "ChronoEngine");
+        AddStaticLibrary( "ChronoEngine_vehicle");
+        AddStaticLibrary( "ChronoModels_vehicle");
+        AddStaticLibrary( "ChronoModels_robot");
 
         bUseRTTI = true;
       }
 
       if (UsingPytorch)
       {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_pytorch")));
+        AddStaticLibrary( "carla_pytorch");
 
-        string LibTorchPath = LibCarlaInstallPath;
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libonnx_proto.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libfbgemm.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgloo.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libXNNPACK.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libprotobuf-lite.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libprotobuf.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libasmjit.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libcpuinfo_internals.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libclog.a"));
-        // PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libbreakpad_common.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libbenchmark.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libtensorpipe.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libpytorch_qnnpack.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libtensorpipe_cuda.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libnnpack_reference_layers.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgmock.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libdnnl.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libpthreadpool.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libcpuinfo.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libqnnpack.a"));
-        // PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libbreakpad.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libkineto.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libprotoc.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgtest.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgmock_main.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgtest_main.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libbenchmark_main.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libfmt.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libtensorpipe_uv.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libfoxi_loader.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libgloo_cuda.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libnnpack.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libcaffe2_protos.a"));
-        PublicAdditionalLibraries.Add(Path.Combine(LibTorchPath, "lib", "libonnx.a"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libnnapi_backend.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libbackend_with_compiler.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libcaffe2_nvrtc.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch_cuda_cpp.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libc10_cuda.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorchbind_test.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libjitbackend_test.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libc10.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch_cuda.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch_global_deps.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch_cpu.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libshm.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorch_cuda_cu.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorchscatter.so"));
-        AddDynamicLibrary(Path.Combine(LibTorchPath, "lib", "libtorchcluster.so"));
-        // AddDynamicLibrary("/usr/local/cuda/lib64/stubs/libcuda.so");
-        // AddDynamicLibrary("/usr/local/cuda/lib64/libnvrtc.so");
-        // AddDynamicLibrary("/usr/local/cuda/lib64/libnvToolsExt.so");
-        // AddDynamicLibrary("/usr/local/cuda/lib64/libcudart.so");
-        // AddDynamicLibrary("/usr/lib/llvm-10/lib/libgomp.so");
-        PublicAdditionalLibraries.Add("/usr/local/cuda/lib64/stubs/libcuda.so");
-        PublicAdditionalLibraries.Add("/usr/local/cuda/lib64/libnvrtc.so");
-        PublicAdditionalLibraries.Add("/usr/local/cuda/lib64/libnvToolsExt.so");
-        PublicAdditionalLibraries.Add("/usr/local/cuda/lib64/libcudart.so");
-        PublicAdditionalLibraries.Add("/usr/lib/llvm-10/lib/libgomp.so");
-        RuntimeDependencies.Add(Path.Combine(LibTorchPath, "lib", "libcudart-a7b20f20.so.11.0"));
-        RuntimeDependencies.Add(Path.Combine(LibTorchPath, "lib", "libgomp-a34b3233.so.1"));
-        RuntimeDependencies.Add(Path.Combine(LibTorchPath, "lib", "libnvrtc-builtins-4730a239.so.11.3"));
-        RuntimeDependencies.Add(Path.Combine(LibTorchPath, "lib", "libnvrtc-1ea278b5.so.11.2"));
-        RuntimeDependencies.Add(Path.Combine(LibTorchPath, "lib", "libnvToolsExt-24de1d56.so.1"));
+        AddStaticLibrary("onnx_proto");
+        AddStaticLibrary("fbgemm");
+        AddStaticLibrary("gloo");
+        AddStaticLibrary("XNNPACK");
+        AddStaticLibrary("protobuf-lite");
+        AddStaticLibrary("protobuf");
+        AddStaticLibrary("asmjit");
+        AddStaticLibrary("cpuinfo_internals");
+        AddStaticLibrary("clog");
+        // AddStaticLibrary("breakpad_common");
+        AddStaticLibrary("benchmark");
+        AddStaticLibrary("tensorpipe");
+        AddStaticLibrary("pytorch_qnnpack");
+        AddStaticLibrary("tensorpipe_cuda");
+        AddStaticLibrary("nnpack_reference_layers");
+        AddStaticLibrary("gmock");
+        AddStaticLibrary("dnnl");
+        AddStaticLibrary("pthreadpool");
+        AddStaticLibrary("cpuinfo");
+        AddStaticLibrary("qnnpack");
+        // AddStaticLibrary("breakpad"));
+        AddStaticLibrary("kineto");
+        AddStaticLibrary("protoc");
+        AddStaticLibrary("gtest");
+        AddStaticLibrary("gmock_main");
+        AddStaticLibrary("gtest_main");
+        AddStaticLibrary("benchmark_main");
+        AddStaticLibrary("fmt");
+        AddStaticLibrary("tensorpipe_uv");
+        AddStaticLibrary("foxi_loader");
+        AddStaticLibrary("gloo_cuda");
+        AddStaticLibrary("nnpack");
+        AddStaticLibrary("caffe2_protos");
+        AddStaticLibrary("onnx");
+        AddDynamicLibrary("torch");
+        AddDynamicLibrary("nnapi_backend");
+        AddDynamicLibrary("backend_with_compiler");
+        AddDynamicLibrary("caffe2_nvrtc");
+        AddDynamicLibrary("torch_cuda_cpp");
+        AddDynamicLibrary("c10_cuda");
+        AddDynamicLibrary("torchbind_test");
+        AddDynamicLibrary("jitbackend_test");
+        AddDynamicLibrary("c10");
+        AddDynamicLibrary("torch_cuda");
+        AddDynamicLibrary("torch_global_deps");
+        AddDynamicLibrary("torch_cpu");
+        AddDynamicLibrary("shm");
+        AddDynamicLibrary("torch_cuda_cu");
+        AddDynamicLibrary("torchscatter");
+        AddDynamicLibrary("torchcluster");
+        AddDynamicLibrary("cuda", "/usr/local/cuda/lib64/stubs");
+        AddDynamicLibrary("nvrtc", "/usr/local/cuda/lib64");
+        AddDynamicLibrary("nvToolsExt", "/usr/local/cuda/lib64");
+        AddDynamicLibrary("cudart", "/usr/local/cuda/lib64");
+        AddDynamicLibrary("gomp", "/usr/lib/llvm-10/lib");
+        RuntimeDependencies.Add(Path.Combine("libcudart-a7b20f20.so.11.0"));
+        RuntimeDependencies.Add(Path.Combine("libgomp-a34b3233.so.1"));
+        RuntimeDependencies.Add(Path.Combine("libnvrtc-builtins-4730a239.so.11.3"));
+        RuntimeDependencies.Add(Path.Combine("libnvrtc-1ea278b5.so.11.2"));
+        RuntimeDependencies.Add(Path.Combine("libnvToolsExt-24de1d56.so.1"));
         PublicAdditionalLibraries.Add("stdc++");
         PublicAdditionalLibraries.Add("/usr/lib/x86_64-linux-gnu/libpython3.9.so");
       }
-
-      if (UsingRos2)
-      {
-        PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", GetLibName("carla_fastdds")));
-
-        string LibFastDDSPath = LibCarlaInstallPath;
-        AddDynamicLibrary(Path.Combine(LibFastDDSPath, "lib", "libfoonathan_memory-0.7.3.so"));
-        AddDynamicLibrary(Path.Combine(LibFastDDSPath, "lib", "libfastcdr.so"));
-        AddDynamicLibrary(Path.Combine(LibFastDDSPath, "lib", "libfastrtps.so"));
-        PublicAdditionalLibraries.Add("stdc++");
-      }
-
-
-      //OsmToODR
-      PublicAdditionalLibraries.Add("/usr/lib/x86_64-linux-gnu/libc.so");
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "libsqlite3.so"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "libxerces-c.a"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "libproj.a"));
-      PublicAdditionalLibraries.Add(Path.Combine(LibCarlaInstallPath, "lib", "libosm2odr.a"));
-
     }
-    bEnableExceptions = true;
+
+    addOsmToODR();
+
+    addROS2();
+
 
     // Include path.
     string LibCarlaIncludePath = Path.Combine(LibCarlaInstallPath, "include");
-
     PublicIncludePaths.Add(LibCarlaIncludePath);
     PrivateIncludePaths.Add(LibCarlaIncludePath);