+++ Util/BuildTools/BuildLibCarla.sh
@@ -125,28 +125,31 @@ function build_libcarla {
 
   CMAKE_EXTRA_OPTIONS=''
 
+  M_ROS=false
   if [ $1 == Server ] ; then
-    M_TOOLCHAIN=${LIBCPP_TOOLCHAIN_FILE}
+    M_TOOLCHAIN=${CARLA_SERVER_TOOLCHAIN_FILE}
     M_BUILD_FOLDER=${LIBCARLA_BUILD_SERVER_FOLDER}.$(echo "$2" | tr '[:upper:]' '[:lower:]')
     M_INSTALL_FOLDER=${LIBCARLA_INSTALL_SERVER_FOLDER}
+  elif [ $1 == ServerROS ] ; then
+    BUILD_TYPE='Server'
+    M_TOOLCHAIN=${CARLA_SERVER_TOOLCHAIN_FILE}
+    M_BUILD_FOLDER=${LIBCARLA_BUILD_SERVER_FOLDER}.ROS.$(echo "$2" | tr '[:upper:]' '[:lower:]')
+    M_INSTALL_FOLDER=${LIBCARLA_INSTALL_SERVER_FOLDER}
+    M_ROS=true
   elif [ $1 == Client ] ; then
-    M_TOOLCHAIN=${LIBSTDCPP_TOOLCHAIN_FILE}
+    M_TOOLCHAIN=${CARLA_CLIENT_TOOLCHAIN_FILE}
     M_BUILD_FOLDER=${LIBCARLA_BUILD_CLIENT_FOLDER}.$(echo "$2" | tr '[:upper:]' '[:lower:]')
     M_INSTALL_FOLDER=${LIBCARLA_INSTALL_CLIENT_FOLDER}
   elif [ $1 == Pytorch ] ; then
-    M_TOOLCHAIN=${LIBSTDCPP_TOOLCHAIN_FILE}
+    M_TOOLCHAIN=${CARLA_SERVER_TOOLCHAIN_FILE}
     M_BUILD_FOLDER=${LIBCARLA_BUILD_PYTORCH_FOLDER}.$(echo "$2" | tr '[:upper:]' '[:lower:]')
-    M_INSTALL_FOLDER=${LIBCARLA_INSTALL_SERVER_FOLDER}
-  elif [ $1 == ros2 ] ; then
-    M_TOOLCHAIN=${LIBSTDCPP_TOOLCHAIN_FILE}
-    M_BUILD_FOLDER=${LIBCARLA_FASTDDS_FOLDER}.$(echo "$2" | tr '[:upper:]' '[:lower:]')
     M_INSTALL_FOLDER=${LIBCARLA_INSTALL_SERVER_FOLDER}
   elif [ $1 == ClientRSS ] ; then
     BUILD_TYPE='Client'
-    M_TOOLCHAIN=${LIBSTDCPP_TOOLCHAIN_FILE}
+    M_TOOLCHAIN=${CARLA_CLIENT_TOOLCHAIN_FILE}
     M_BUILD_FOLDER=${LIBCARLA_BUILD_CLIENT_FOLDER}.rss.$(echo "$2" | tr '[:upper:]' '[:lower:]')
     M_INSTALL_FOLDER=${LIBCARLA_INSTALL_CLIENT_FOLDER}
-    CMAKE_EXTRA_OPTIONS="${CMAKE_EXTRA_OPTIONS:+${CMAKE_EXTRA_OPTIONS} }-DBUILD_RSS_VARIANT=ON -DADRSS_INSTALL_DIR=${CARLA_BUILD_FOLDER}/ad-rss-4.4.4/install"
+    CMAKE_EXTRA_OPTIONS="${CMAKE_EXTRA_OPTIONS:+${CMAKE_EXTRA_OPTIONS} }-DBUILD_RSS_VARIANT=ON -DADRSS_INSTALL_DIR=${CARLA_BUILD_FOLDER}/ad-rss-4.4.5/install"
   else
     fatal_error "Invalid build configuration \"$1\""
   fi
@@ -184,6 +187,7 @@ function build_libcarla {
         -DCMAKE_BUILD_TYPE=${BUILD_TYPE:-$1} \
         -DLIBCARLA_BUILD_DEBUG=${M_DEBUG} \
         -DLIBCARLA_BUILD_RELEASE=${M_RELEASE} \
+        -DLIBCARLA_USE_ROS=${M_ROS} \
         -DCMAKE_TOOLCHAIN_FILE=${M_TOOLCHAIN} \
         -DCMAKE_INSTALL_PREFIX=${M_INSTALL_FOLDER} \
         -DCMAKE_EXPORT_COMPILE_COMMANDS=1 \
@@ -194,7 +198,7 @@ function build_libcarla {
 
   fi
 
-  ninja
+  ninja 
 
   ninja install | grep -v "Up-to-date:"
 
@@ -205,23 +209,22 @@ function build_libcarla {
 # -- Build all possible configurations -----------------------------------------
 # ==============================================================================
 
+SERVER_VARIANT='Server'
+if ${USE_ROS2}; then
+  SERVER_VARIANT='ServerROS'
+fi
 if { ${BUILD_SERVER} && ${BUILD_OPTION_DEBUG}; }; then
 
-  build_libcarla Server Debug
+  build_libcarla ${SERVER_VARIANT} Debug
 
 fi
 
 if { ${BUILD_SERVER} && ${BUILD_OPTION_RELEASE}; }; then
 
-  build_libcarla Server Release
+  build_libcarla ${SERVER_VARIANT} Release
   if ${USE_PYTORCH} ; then
     build_libcarla Pytorch Release
   fi
-
-  if ${USE_ROS2} ; then
-    build_libcarla ros2 Release
-  fi
-
 fi
 
 CLIENT_VARIANT='Client'